Fixes to correct app crash during device orientation change
Added android:configChanges attribute to <activity> to fix
application crashes if the user changes the phone's orientation
while the progress dialog is not yet dismissed.

Activities in Android are destroyed and recreated on an orientation
change. Because the progress dialog is no longer "attached" to the
view after the activity is re-created, it blows up when an attempt
is made to reference it.