Remove references to Small Traffic site.
In the Traffic & Travel section we provided a link to the small traffic
site. This link was only meant to stay until we could replicate the
functionality within the app. With the addition of Ferry alerts and
schedules most of content from the small traffic site is in the app.