Add new preference to show shadows behind markers on map.
According to the docs the marker will be drawn twice for each Item in
the Overlay - once in the shadow phase, skewed and darkened, then
again in the non-shadow phase.

Not drawing the shadow means 1/2 as many markers on the map and
performance improves for those with older phones like the G1.