Dynamically add and remove Seattle area specific menu items
Added a geo fence around the Seattle area in which Seattle
travel times and alerts should show in the menu if the map center
point is within its boundaries. If the map center is outside the
boundary the menu items are removed.

No sense showing Seattle travel times to someone in Eastern WA.
its boundaries.