Add setOnCancelListener to GetCameraImage AsyncTask class
If the camera image is taking to long to download and the user
presses the back button to cancel the dialog the thread is still
running in the background and the image will sometimes pop up later.
Sometimes a crash will happen.

Setting the setOnCancelListener should take care of this.