Optimize ListView layout_height property
At Google IO 2010 there was a presentation by Romain Guy and Adam Powell
talking about ListView's layout_height. Using wrap_content to define the
ListView's height creates a tremendous amount of overhead for ones
application. Their example of the proper way to layout a ListView is to
set the layout_height to "0dp" and layout_weight to "1.0"

http://code.google.com/events/io/2010/sessions/world-of-listview-android.html