Initial commit with support for Android 4.0.3 (Ice Cream Sandwich)
Using ActionBarSherlock library. ActionBarSherlock is an extension
of the compatibility library designed to facilitate the use of the
action bar design pattern across all versions of Android with a
single API.

The library will automatically use the native action
bar when appropriate or will automatically wrap a custom
implementation around your layouts. This allows you to easily
develop an application with an action bar for every version of
Android from 2.x and up.

More information: http://actionbarsherlock.com/