Changing menu items at runtime - fix for Android 3.0+
On Android 2.3.x and lower, the system calls onPrepareOptionsMenu() each
time the user opens the options menu (presses the Menu button).

On Android 3.0 and higher, the options menu is considered to always be open
when menu items are presented in the action bar. When an event occurs and
you want to perform a menu update, you must call invalidateOptionsMenu()
to request that the system call onPrepareOptionsMenu().

http://developer.android.com/guide/topics/ui/menus.html#ChangingTheMenu