Implement swiping across mountain pass tabs.
Removed the gallery view widget as the horizontal swiping conflicts
with it. Replaced it with a vertically scrolling list view where the
camera images are resized to the width of the view and maintain their
correct aspect ratio. Can't hard code the camera image sizes because
we have a number of different sizes.

Images are correctly resized with a custom class which subclasses Android's
ImageView. See http://stackoverflow.com/a/6046532 for details.

Mountain pass cameras can also be selected from the listview and shared,
favorited or refreshed.