Request permissions at runtime for ACCESS_FINE_LOCATION
Beginning in Android 6.0 (API level 23), users grant permissions
to apps while the app is running, not when they install the app.

This commit checks if we access to use the GPS before making the
call to Location Services. There is an issue that if the user
checks "Never ask again" on the permission request dialog the
onConnected method continues to repeatedly hammer calls requesting
permission. Not sure how to code around this.