Fix for users reporting "Can't install app" error message from Play Store
The signature algorithms for jarsigner changed in JDK 7. The default
algorithms have been bumped to SHA-256 and SHA256withRSA respectively,
and are not supported on Android.

Based on this answer from stackoverflow, http://stackoverflow.com/a/9567153
and the Android developer documentation on how to sign your app
manually, http://developer.android.com/tools/publishing/app-signing.html#signing-manually
the additional arguments were added to the mavin-jarsigner-plugin
configuration to override the default signature algorithms.

Closes #20