Added alert box for when FCM is received while in foreground
FCM messages are handled in one of two ways.
1. App is in foreground and user is in any activity.
The app will display and alert and ask the user if they would like to
be directed to an activity pertaining to the alert if data is included
in the payload.
2. App is in the background.
App will open to the activity if it was included. Otherwise it will
just open to the HomeActivity.