Added refresh to pass report screen
The MountainPassItemActivity now handles refreshing the data in it's
managed tabs. When the user pressed the refresh button the corresponding
refresh method is called in the displayed fragment and the data is
reloaded.

To achieve this, the report and forecast fragment now only receive the
pass ID and fetch their content from the database.

Pass cameras get the time since 1970 tacked onto the end of the URL to
force the ImageManager to reload the image.