Fixes refresh indicator in favorites section
Changes when NetworkResourceSyncRepository updates the loading status.
Now sets the status before starting thread work to ensure the new
status is delivered to the view before it completes.
Removes unnecessary thread work in ferries section, this is also
required for the above change, since setValue can’t be called on a
thread.