Replaces EventService with two new Worker classes
Worker classes from the WorkManger package now handle
checking for event banners and new subscription topics

The EventService just makes two network calls and updates
Shared Preferences. Making it a Service was overkill and lead to issues when the app would enter the background after the service has started.

Two new workers, EventWorker (to check for active banner text)
and TopicWorker (to check for new notification topics) replace the old service.

An added bonus is that the event banner will now appear as soon
as the worker completes.